requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.8.2
cachetools==5.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docker==7.1.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
Flask==3.0.3
fonttools==4.38.0
google-auth==2.32.0
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.1
jupyter_client==7.4.5
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
kubernetes==30.1.0
lldb-python==19.0.0.dev1
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
oauthlib==3.2.2
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.3
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
Pygments==2.13.0
pymongo==4.9.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.1.0
Send2Trash==1.8.0
setuptools @ file:///opt/homebrew/Cellar/python%403.10/3.10.14_1/Frameworks/Python.framework/Versions/3.10/lib/python3.10/ensurepip/_bundled/setuptools-70.3.0-py3-none-any.whl#sha256=22463f85743c9278393c7fbe2575b2760bda74a1a91a4183347ba444b2e6f876
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.0
terminado==0.17.0
tinycss2==1.2.1
tornado==6.2
traitlets==5.5.0
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==3.0.4
wheel @ file:///opt/homebrew/Cellar/python%403.10/3.10.14_1/libexec/wheel-0.43.0-py3-none-any.whl#sha256=293b2a06aa87c255cc286400027c0152575b319877e07ff9542590383532acfc
widgetsnbextension==4.0.3
